requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
feedparser>=6.0.10

# Technical Analysis & AI
talib
scikit-learn
ta

# HTTP client enhancements
urllib3>=1.26.0

# Additional ML/NLP dependencies
tokenizers>=0.13.0
safetensors>=0.3.0

# Development and testing (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Logging enhancements
colorlog>=6.7.0

# Telegram Bot
python-telegram-bot>=21.0